💄 dark mode support
Normally when implementing a client side dark mode in a SSR-enabled app, there will be a brief white flash while loading the page before the main app's script executes.

This is fixed by having a small script to patch the style sheet at the beginning. I have implemented a simple patching script in this project. It works but there are some caveats:
- It requires the server to render twice. This isn't a problem because we use static builds.
- You can only change palette color in the theme. You can't apply different props based on light/dark theme.